Prevent writing multiple log entries
Due to an error in SCM-Manager up to version 2.39.0
the changeset provider returned all changesets for a git
repository reachable by a branch, if this branch was pushed
without a new commit. This led to redundant pushlog entries
for commits.

With this change we check whether an announced commit id already
has a pushlog entry. If this is the case, the further process
will be aborted.