Prevent deadlock for pushlog events
If a repository has been deleted before pushlogs have been written, it
could happen that the event bus handling the post receive hooks runs
into a deadlock. This is because the data store keeping the pushlogs
throws an exception, when the repository is no longer available.
Therefore no pushlog entity will be created and the lock that had been
created will not be unlocked. If in this case a second pushlog for this
repository shall be written (because two pushes had been made before
the repository had been deleted), it cannot pass the lock and therefore
will block the eventbus forever, preventing any other pushlog hooks from
ever being processed.

To prevent this, we handle locking and unlocking in a single function
with an overall try/finally method.